extract profile form checking logic
